+++ Application/Feature/Rol/Commands/ActualizarRolCommand.cs
@@ -7,7 +7,7 @@ namespace Application.Feature.Rol.Commands
 {
     public class ActualizarRolCommand : IRequest<ObjectResponse<RolResponse>>
     {
-        public int pId { get; set; }
+        public long pId { get; set; }
         public string pNombre { get; set; }
         public string pDescripcion { get; set; }
     }
